+++ test/tbb/test_concurrent_queue.cpp
@@ -1,5 +1,5 @@
 /*
-    Copyright (c) 2005-2023 Intel Corporation
+    Copyright (c) 2005-2022 Intel Corporation
 
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
@@ -18,7 +18,6 @@
 #include <common/utils.h>
 #include <common/vector_types.h>
 #include <common/custom_allocators.h>
-#include <common/container_move_support.h>
 
 #include <tbb/concurrent_queue.h>
 #include <unordered_set>
@@ -281,212 +280,4 @@ void test_tracking_dtors_on_clear() {
 TEST_CASE("Test clear and dtor with TrackableItem") {
     test_tracking_dtors_on_clear<oneapi::tbb::concurrent_queue<TrackableItem>>();
     test_tracking_dtors_on_clear<oneapi::tbb::concurrent_bounded_queue<TrackableItem>>();
-}
-
-template<typename CQ, typename T>
-void test_queue_helper() {
-    int size = 5;
-    T vec_1(size, 0), vec_2(size, 0), vec_3(size, 0), vec_4(size, 0);
-    srand(static_cast<unsigned>(time(0)));
-    generate(vec_1.begin(), vec_1.end(), rand);
-    generate(vec_2.begin(), vec_2.end(), rand);
-    generate(vec_3.begin(), vec_3.end(), rand);
-    generate(vec_4.begin(), vec_4.end(), rand);
-
-    CQ q1, q2, q3;
-    CQ q4({vec_1, vec_2, vec_3});
-    q3 = {vec_4, vec_2, vec_3};
-
-    q1 = q3;
-    q2 = std::move(q3);
-    CHECK(q3.empty());
-
-    CHECK(q1 != q4);
-    q1.swap(q4);
-    CHECK(q2 == q4);
-
-    swap(q2, q3);
-    CHECK(q2.empty());
-    CHECK(q3 == q4);
-}
-
-//! basic test for copy, move and swap
-//! \brief \ref error_guessing
-TEST_CASE("Test iterator queue") {
-    test_queue_helper<tbb::concurrent_queue<std::vector<int>>, std::vector<int>>();
-    test_queue_helper<tbb::concurrent_bounded_queue<std::vector<int>>, std::vector<int>>();
-}
-
-template <typename queue_type, typename allocator_type>
-void TestMoveQueue() {
-    allocator_type::set_limits(300);
-    queue_type q1, q2;
-    move_support_tests::Foo obj;
-    size_t n1(15), n2(7);
-
-    allocator_type::init_counters();
-    for(size_t i =0; i < n1; i++)
-        q1.push(obj);
-    size_t q1_items_constructed = allocator_type::items_constructed;
-    size_t q1_items_allocated =  allocator_type::items_allocated;
-
-    allocator_type::init_counters();
-    for(size_t i =0; i < n2; i++)
-        q2.push(obj);
-    size_t q2_items_allocated =  allocator_type::items_allocated;
-
-    allocator_type::init_counters();
-    q1 = std::move(q2);
-
-    CHECK(q1_items_allocated == allocator_type::items_freed);
-    CHECK(q1_items_constructed == allocator_type::items_destroyed);
-    CHECK(q2_items_allocated >= allocator_type::items_allocated);
-}
-
-//! move assignment test for equal allocator
-//! \brief \ref error_guessing
-TEST_CASE("Test move queue") {
-    using allocator_type = StaticSharedCountingAllocator<std::allocator<move_support_tests::Foo>>;
-    TestMoveQueue<tbb::concurrent_queue<move_support_tests::Foo, allocator_type>, allocator_type>();
-    TestMoveQueue<tbb::concurrent_bounded_queue<move_support_tests::Foo, allocator_type>, allocator_type>();
-}
-
-template<class T>
-struct stateful_allocator {
-    typedef T value_type;
-    stateful_allocator () = default;
-    int state = 0;
-    template<class U>
-    constexpr stateful_allocator (const stateful_allocator <U>&) noexcept {}
-
-    T* allocate(std::size_t n) {
-        return static_cast<T*>(std::malloc(n * sizeof(T)));
-    }
-
-    void deallocate(T* p, std::size_t n) noexcept {
-        static_cast<void>(n);
-        std::free(p);
-    }
-};
-
-template<class T, class U>
-bool operator==(const stateful_allocator <T>& lhs, const stateful_allocator <U>& rhs) { return &lhs.state == &rhs.state; }
-
-template<class T, class U>
-bool operator!=(const stateful_allocator <T>& lhs, const stateful_allocator <U>& rhs) { return &lhs.state != &rhs.state; }
-
-template <typename queue_type, typename allocator_type>
-void TestMoveQueueUnequal() {
-    allocator_type::set_limits(300);
-    queue_type q1, q2;
-    move_support_tests::Foo obj;
-    size_t n1(15), n2(7);
-
-    allocator_type::init_counters();
-    for(size_t i =0; i < n1; i++)
-        q1.push(obj);
-
-    allocator_type::init_counters();
-    for(size_t i =0; i < n2; i++)
-        q2.push(obj);
-    size_t q2_items_allocated =  allocator_type::items_allocated;
-
-    allocator_type::init_counters();
-    q1 = std::move(q2);
-
-    REQUIRE_MESSAGE( (allocator_type::items_allocated - q2_items_allocated) <= 0, "More then excepted memory allocated ?" );
-    REQUIRE_MESSAGE(std::all_of(q1.unsafe_begin(), q1.unsafe_end(), is_state_predicate<move_support_tests::Foo::MoveInitialized>()),
-                    "Container did not move construct some elements");
-    REQUIRE_MESSAGE(std::all_of(q2.unsafe_begin(), q2.unsafe_end(), is_state_predicate<move_support_tests::Foo::MovedFrom>()),
-                    "Container did not move all the elements");
-}
-
-//! move assignment test for unequal allocator
-//! \brief \ref error_guessing
-TEST_CASE("Test move queue-unequal allocator") {
-    using allocator_type = StaticSharedCountingAllocator<stateful_allocator<move_support_tests::Foo>>;
-    TestMoveQueueUnequal<tbb::concurrent_queue<move_support_tests::Foo, allocator_type>, allocator_type>();
-    TestMoveQueueUnequal<tbb::concurrent_bounded_queue<move_support_tests::Foo, allocator_type>, allocator_type>();
-}
-
-template<typename Container>
-void test_check_move_allocator(Container& src, Container& dst, Container& cpy) {
-    if(src.get_allocator() == dst.get_allocator()){
-        REQUIRE_MESSAGE(src.empty(), "Source didn't clear");
-        REQUIRE_MESSAGE(std::equal(dst.unsafe_begin(), dst.unsafe_end(), cpy.unsafe_begin()), "Elements are not equal");
-    } else {
-        REQUIRE_MESSAGE(&*(src.unsafe_begin()) !=  &*(dst.unsafe_begin()), "Container did not change element locations for unequal allocators");
-        REQUIRE_MESSAGE(std::equal(dst.unsafe_begin(), dst.unsafe_end(), cpy.unsafe_begin()), "Elements are not equal");
-    }
-}
-
-void test_move_assignment_test_equal() {
-    int n = 5;
-    std::vector<int> vect1(n, 10), vect2(n,20), vect3(n, 30);
-
-    tbb::concurrent_queue<std::vector<int>> src({vect1, vect2, vect3});
-    tbb::concurrent_queue<std::vector<int>> dst(src.get_allocator());
-    tbb::concurrent_queue<std::vector<int>> cpy(src.get_allocator());
-    cpy = src;
-    dst = std::move(src);
-
-    tbb::concurrent_bounded_queue<std::vector<int>> src_bnd({vect1, vect2, vect3});
-    tbb::concurrent_bounded_queue<std::vector<int>> dst_bnd(src_bnd.get_allocator());
-    tbb::concurrent_bounded_queue<std::vector<int>> cpy_bnd(src_bnd.get_allocator());
-    cpy_bnd = src_bnd;
-    dst_bnd = std::move(src_bnd);
-
-    test_check_move_allocator<tbb::concurrent_queue<std::vector<int>>>(src, dst, cpy);
-    REQUIRE_MESSAGE(src.get_allocator() == dst.get_allocator(), "Incorrect test setup: allocators should be equal");
-    REQUIRE_MESSAGE(cpy.unsafe_size() == dst.unsafe_size(), "Queues are not equal");
-
-    test_check_move_allocator<tbb::concurrent_bounded_queue<std::vector<int>>>(src_bnd, dst_bnd, cpy_bnd);
-    REQUIRE_MESSAGE(src.get_allocator() == dst.get_allocator(), "Incorrect test setup: allocators should be equal");
-    REQUIRE_MESSAGE(cpy_bnd.size() == dst_bnd.size(), "Queues are not equal");
-}
-
-void test_move_assignment_test_unequal() {
-    stateful_allocator<int> src_alloc;
-    std::vector<int, stateful_allocator<int>> v(8, 0, src_alloc);
-    tbb::concurrent_queue<std::vector<int, stateful_allocator<int>>, stateful_allocator<int>> src(src_alloc);
-
-    src_alloc.state = 0;
-    v.push_back(42);
-    v.push_back(82);
-    src.push(v);
-    src.push(v);
-
-    stateful_allocator<int> dst_alloc;
-    dst_alloc.state = 1;
-    tbb::concurrent_queue<std::vector<int, stateful_allocator<int>>, stateful_allocator<int>> dst(dst_alloc);
-    tbb::concurrent_queue<std::vector<int, stateful_allocator<int>>, stateful_allocator<int>> cpy(src_alloc);
-    cpy = src;
-    dst = std::move(src);
-
-    tbb::concurrent_bounded_queue<std::vector<int, stateful_allocator<int>>, stateful_allocator<int>> src_bnd(src_alloc);
-    tbb::concurrent_bounded_queue<std::vector<int, stateful_allocator<int>>, stateful_allocator<int>> dst_bnd(dst_alloc);
-    tbb::concurrent_bounded_queue<std::vector<int, stateful_allocator<int>>, stateful_allocator<int>> cpy_bnd(src_alloc);
-    src_bnd.push(v);
-    src_bnd.push(v);
-    cpy_bnd = src_bnd;
-    dst_bnd = std::move(src_bnd);
-
-    test_check_move_allocator<tbb::concurrent_queue<std::vector<int, stateful_allocator<int>>, stateful_allocator<int>>>(src, dst, cpy);
-    REQUIRE_MESSAGE(src.get_allocator() != dst.get_allocator(), "Incorrect test setup: allocators should be unequal");
-    REQUIRE_MESSAGE(src.unsafe_size() == 0, "Moved from container should not contain any elements");
-    REQUIRE_MESSAGE(dst.unsafe_size() == cpy.unsafe_size(), "Queues are not equal");
-    REQUIRE_MESSAGE(std::equal(dst.unsafe_begin(), dst.unsafe_end(), cpy.unsafe_begin()), "Elements are not equal");
-
-    test_check_move_allocator<tbb::concurrent_bounded_queue<std::vector<int, stateful_allocator<int>>, stateful_allocator<int>>>(src_bnd, dst_bnd, cpy_bnd);
-    REQUIRE_MESSAGE(src.get_allocator() != dst.get_allocator(), "Incorrect test setup: allocators should be unequal");
-    REQUIRE_MESSAGE(src_bnd.size() == 0, "Moved from container should not contain any elements");
-    REQUIRE_MESSAGE(dst_bnd.size() == cpy_bnd.size(), "Queues are not equal");
-    REQUIRE_MESSAGE(std::equal(dst_bnd.unsafe_begin(), dst_bnd.unsafe_end(), cpy_bnd.unsafe_begin()), "Elements are not equal");
-}
-
-//! move assignment test for equal and unequal allocator
-//! \brief \ref error_guessing
-TEST_CASE("concurrent_queue") {
-    test_move_assignment_test_equal();
-    test_move_assignment_test_unequal();
 }
